Do not make Theano cache code delete the lock directory. This fix problems happening in the daily buildbot due to concurrent Theano process.

+++ theano/gof/cmodule.py
@@ -673,6 +673,10 @@ class ModuleCache(object):
                 continue
             if not os.path.isdir(root):
                 continue
+            # Some sub directory we do not want the cache to mess
+            # with.  This can cause problems with multiple process.
+            if os.path.split(root)[1] in ["lock_dir"]:
+                continue
             files = os.listdir(root)
             if not files or 'delete.me' in files:
                 rmtree(root, ignore_nocleanup=True,
